# callcenter.tj-api/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@callcentrix.com
    name: API Support
  description: API for call center monitoring and control
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CallCentrix API
  version: "1.0"
paths:
  /api/actions/hangup:
    post:
      consumes:
      - application/json
      description: Force hangup call by callId (Linkedid)
      parameters:
      - description: Call Linkedid
        in: query
        name: callId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: missing callId
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: call not found
          schema:
            type: string
        "500":
          description: AMI not available
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Hangup call
      tags:
      - Actions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"